requirements.txt
Flask==3.0.2
Faker==23.2.1
requests==2.31.0
Pillow==10.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pycryptodome==3.20.0
psutil==5.9.8
plyer==2.1.0
torch==2.3.0
numpy==1.26.4
matplotlib==3.8.4
pandas==2.2.2
tqdm==4.66.2
wordcloud==1.9.3
openpyxl==3.1.2